Fjerner ny utbetalingsgenerator som ikke virket som forventet og ikke tatt i bruk. Vi kommer å skrive en ny som håndterer kjeder på en oppdatert måte.
