+++ src/test/enhetstester/kotlin/no/nav/familie/ba/sak/integrasjoner/økonomi/ØkonomiTestUtils.kt
@@ -1,6 +1,10 @@
 package no.nav.familie.ba.sak.integrasjoner.økonomi
 
 import no.nav.familie.ba.sak.kjerne.beregning.domene.YtelseType
+import no.nav.familie.kontrakter.felles.oppdrag.Utbetalingsoppdrag
+import no.nav.familie.kontrakter.felles.oppdrag.Utbetalingsperiode
+import java.time.LocalDateTime
+import java.util.UUID
 
 fun sats(ytelseType: YtelseType) =
     when (ytelseType) {
@@ -9,3 +13,13 @@ fun sats(ytelseType: YtelseType) =
         YtelseType.SMÅBARNSTILLEGG -> 660
         YtelseType.MANUELL_VURDERING -> 0
     }
+
+fun lagUtbetalingsoppdrag(utbetalingsperiode: List<Utbetalingsperiode>) = Utbetalingsoppdrag(
+    kodeEndring = Utbetalingsoppdrag.KodeEndring.NY,
+    fagSystem = "BA",
+    saksnummer = "",
+    aktoer = UUID.randomUUID().toString(),
+    saksbehandlerId = "",
+    avstemmingTidspunkt = LocalDateTime.now(),
+    utbetalingsperiode = utbetalingsperiode
+)
